refactor: Separate data endpoints from legacy search endpoints
[#OCD-3601]
